+++ src/main/java/org/markdownwriterfx/dialogs/LinkDialog.java
@@ -32,7 +32,6 @@ import javafx.application.Platform;
 import javafx.beans.binding.Bindings;
 import javafx.beans.property.SimpleStringProperty;
 import javafx.beans.property.StringProperty;
-import javafx.scene.control.ButtonBar.ButtonData;
 import javafx.scene.control.ButtonType;
 import javafx.scene.control.Dialog;
 import javafx.scene.control.DialogPane;
@@ -78,12 +77,11 @@ public class LinkDialog
 				.then(Bindings.format("[%s](%s \"%s\")", textField.escapedTextProperty(), urlField.escapedTextProperty(), titleField.escapedTextProperty()))
 				.otherwise(Bindings.when(textField.escapedTextProperty().isNotEmpty())
 						.then(Bindings.format("[%s](%s)", textField.escapedTextProperty(), urlField.escapedTextProperty()))
-						.otherwise(urlField.escapedTextProperty())));
+						.otherwise(Bindings.format("<%s>", urlField.escapedTextProperty()))));
 		previewField.textProperty().bind(link);
 
 		setResultConverter(dialogButton -> {
-			ButtonData data = (dialogButton != null) ? dialogButton.getButtonData() : null;
-			return (data == ButtonData.OK_DONE) ? link.get() : null;
+			return (dialogButton == ButtonType.OK) ? link.get() : null;
 		});
 
 		Platform.runLater(() -> {
@@ -92,6 +90,12 @@ public class LinkDialog
 			if (urlField.getText().startsWith("http://"))
 				urlField.selectRange("http://".length(), urlField.getLength());
 		});
+	}
+
+	public void init(String url, String text, String title) {
+		urlField.setText(url);
+		textField.setText(text);
+		titleField.setText(title);
 	}
 
 	private void initComponents() {
@@ -110,7 +114,7 @@ public class LinkDialog
 
 		//======== pane ========
 		{
-			pane.setCols("[shrink 0,fill][300,grow,fill][fill][fill]");
+			pane.setCols("[shrink 0,fill][400,grow,fill]");
 			pane.setRows("[][][][]");
 
 			//---- urlLabel ----
@@ -122,8 +126,14 @@ public class LinkDialog
 			urlField.setText("http://yourlink.com");
 			urlField.setPromptText("http://yourlink.com");
 			pane.add(urlField, "cell 1 0");
-			pane.add(linkBrowseDirectoyButton, "cell 2 0");
-			pane.add(linkBrowseFileButton, "cell 3 0");
+
+			//---- linkBrowseDirectoyButton ----
+			linkBrowseDirectoyButton.setFocusTraversable(false);
+			pane.add(linkBrowseDirectoyButton, "cell 1 0,alignx center,growx 0");
+
+			//---- linkBrowseFileButton ----
+			linkBrowseFileButton.setFocusTraversable(false);
+			pane.add(linkBrowseFileButton, "cell 1 0,alignx center,growx 0");
 
 			//---- textLabel ----
 			textLabel.setText(Messages.get("LinkDialog.textLabel.text"));
@@ -131,17 +141,17 @@ public class LinkDialog
 
 			//---- textField ----
 			textField.setEscapeCharacters("[]");
-			pane.add(textField, "cell 1 1 3 1");
+			pane.add(textField, "cell 1 1");
 
 			//---- titleLabel ----
 			titleLabel.setText(Messages.get("LinkDialog.titleLabel.text"));
 			pane.add(titleLabel, "cell 0 2");
-			pane.add(titleField, "cell 1 2 3 1");
+			pane.add(titleField, "cell 1 2");
 
 			//---- previewLabel ----
 			previewLabel.setText(Messages.get("LinkDialog.previewLabel.text"));
 			pane.add(previewLabel, "cell 0 3");
-			pane.add(previewField, "cell 1 3 3 1");
+			pane.add(previewField, "cell 1 3");
 		}
 		// JFormDesigner - End of component initialization  //GEN-END:initComponents
 	}